+++ examples/provider/variables.tf
@@ -9,3 +9,5 @@ variable "password" {
 variable "validate_certs" {
     type = bool
 }
+
+# Alternate method: For certificate-based auth, see docs/guides/ssl_certificate_authentication.md